Separate individual parsers into their own modules
Introduces a `./lib/parsers` package that houses all individual parsers
for different languages. These may be separated into their own modules
or npm packages in the future.
