+++ lib/cli.js
@@ -3,6 +3,7 @@
 
 var fs         = require('fs');
 var path       = require('path');
+var parsers    = require('./parsers');
 var jsxgettext = require('./jsxgettext');
 
 var opts = require("nomnom")
@@ -45,24 +46,23 @@ var opts = require("nomnom")
   .option('language', {
     abbr: 'L',
     metavar: 'NAME',
-    default: 'JavaScript',
-    help: 'recognise the specified language (JavaScript, EJS, Jinja, Jade, Handlebars)'
+    default: 'javascript',
+    help: 'use the specified language (' + ['javascript'].concat(Object.keys(parsers)).join(', ') + ')'
   })
   .parse();
 
 function gen(sources) {
   var result;
-  if (opts.language.toUpperCase() === 'EJS') {
-    result = jsxgettext.generateFromEJS(sources, opts);
-  } else if (opts.language.toUpperCase() === 'JINJA') {
-    result = jsxgettext.generateFromJinja(sources, opts);
-  } else if (opts.language.toUpperCase() === 'JADE') {
-    result = jsxgettext.generateFromJade(sources, opts);
-  } else if (opts.language.toUpperCase() === 'HANDLEBARS') {
-    result = jsxgettext.generateFromHandlebars(sources, opts);
-  } else {
+  var lang = opts.language.toLowerCase();
+
+  if (lang === 'javascript') {
     result = jsxgettext.generate(sources, opts);
+  } else if (lang in parsers) {
+    result = jsxgettext.generate.apply(jsxgettext, parsers[lang](sources, opts));
+  } else {
+    throw new Error("Unsupported language: " + opts.language);
   }
+
   if (opts.output === '-') {
     console.log(result);
   } else {
