+++ lib/jsxgettext.js
@@ -1,4 +1,3 @@
-/* global require */
 "use strict";
 
 /* This Source Code Form is subject to the terms of the Mozilla Public
@@ -10,7 +9,6 @@ var path = require('path');
 
 var parser        = require('acorn');
 var estraverse    = require('estraverse');
-var jade          = require('jade');
 var gettextParser = require('gettext-parser');
 
 var traverse = estraverse.traverse;
@@ -176,238 +174,4 @@ function gen(sources, options) {
   return gettextParser.po.compile(poJSON).toString();
 }
 
-// strips everything but the javascript bits
-function parseEJS(str, options) {
-  options = options || {};
-  var open = options.open || '<%',
-    close = options.close || '%>';
-
-  var buf = [];
-
-  for (var i = 0, len = str.length; i < len; ++i) {
-    if (str.slice(i, open.length + i) === open) {
-      i += open.length;
-      switch (str.substr(i, 1)) {
-        case '=':
-        case '-':
-          ++i;
-          break;
-      }
-
-      var end = str.indexOf(close, i), js = str.substring(i, end), start = i, n = 0;
-      if ('-' === js[js.length - 1]) {
-        js = js.substring(0, js.length - 2);
-      }
-
-      while ((n = js.indexOf("\n", n)) > -1) {
-        n += 1;
-        buf.push("\n");
-      }
-
-      // skip EJS include statements which are not valid javascript
-      if (/^\s*include\s*[^\s]+\s*$/.test(js)) js = "";
-      buf.push(js, ';');
-      i += end - start + close.length - 1;
-
-    } else if (str.substr(i, 1) === "\n") {
-      buf.push("\n");
-    }
-  }
-
-  return buf.join('');
-}
-
-// generate extracted strings file from EJS
-function genEJS(ejsSources, options) {
-  Object.keys(ejsSources).forEach(function (filename) {
-    ejsSources[filename] = parseEJS(ejsSources[filename]);
-  });
-
-  return gen(ejsSources, options);
-}
-
-// From MDN:
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions#Using_Special_Characters
-function escapeRegExp(string) {
-  return string.replace(/([.*+?^=!:${}()|\[\]\/\\])/g, "\\$1");
-}
-
-function parseJade(str, options) {
-  options = options || {};
-
-  var parser = new jade.Parser(str);
-  var lexer = parser.lexer;
-  var token;
-
-  var gettextRegexPrefix = '(?:gettext';
-  if (options.keyword) {
-    gettextRegexPrefix += '|' + escapeRegExp(options.keyword);
-  }
-  gettextRegexPrefix += ')';
-  var gettexRegex = new RegExp(gettextRegexPrefix + '(?:\\(\"[^"]+\"|\\(\'[^\']+\')', 'gi');
-
-  function extractGettext(str) {
-    if (typeof(str) !== 'string') return '';
-
-    var tmp = str.match(gettexRegex) || [];
-    return tmp.map(function (t) {
-      return t + ')';
-    }).join(';');
-  }
-
-  function extractFromObj(key) {
-    /* jshint -W040 */
-    return extractGettext(this[key]);
-  }
-
-  function isEmpty(obj) {
-    return obj.length;
-  }
-
-  var buf = [], lineN, tmp;
-  do {
-    token = lexer.next();
-    lineN = token.line - 1;
-    switch (token.type) {
-      case 'attrs':
-        tmp = Object.keys(token.attrs).map(extractFromObj, token.attrs).filter(isEmpty);
-        if (tmp.length)
-          buf[lineN] = tmp.join('') + ';';
-        break;
-      case 'text':
-      case 'code':
-        tmp = extractGettext(token.val);
-        if (tmp.length) buf[lineN] = tmp + ';';
-        break;
-    }
-  } while (token.type !== 'eos');
-
-  return buf.join('\n');
-}
-
-// generate extracted strings file from Jade templates
-function genJade(jadeSources, options) {
-  Object.keys(jadeSources).forEach(function (filename) {
-    jadeSources[filename] = parseJade(jadeSources[filename], options);
-  });
-  return gen(jadeSources, options);
-}
-
-// Turn handlebars helper calls into javascript-syntax functions.
-// Also comment blocks are turned into javascript comments.
-function parseHandlebars(str/*, options*/) {
-
-  // Using regexes for parsing, ooooh yeeeahhh!
-  // Short comments:  {{! this is a comment }}
-  var shortCommentRE = /\{\{\!(.*?)\}\}/;
-  // Long comments:  {{!-- this comment has {{markup}} in it --}}
-  var longCommentRE = /\{\{\!--(.*?)--\}\}/;
-  // Block helpers:  {{#helper}}template content{{/helper}}
-  var blockHelperStartRE = /\{\{#(\w+)\}\}/;
-  var blockHelperEndRE = /\{\{\/(\w+)\}\}/;
-  // Function helpers:  {{ helper value }} or {{ helper "some string" }}
-  var singleQuotedStringWithEscapes = "'(([^']*?(\\\\')?)+)'";
-  var doubleQuotedStringWithEscapes = '"(([^"]*?(\\\\")?)+)"';
-  var funcHelperRE = new RegExp("\\{\\{\\s*(\\w+)\\s+((\\w+)|(" +
-                                singleQuotedStringWithEscapes + ")|(" +
-                                doubleQuotedStringWithEscapes + "))\\s*\\}\\}");
-
-  var buf = [];
-  var match = null;
-
-  function addComment(comment) {
-    buf.push("//");
-    buf.push(comment);
-    buf.push("\n");
-  }
-
-  while (str.length) {
-    // Find the earliest match of any type of tag in the string.
-    match = str.match(shortCommentRE);
-    if (match) {
-      match.type = 'comment';
-    }
-    var nextMatch = str.match(longCommentRE);
-    if (nextMatch) {
-      if (!match || nextMatch.index < match.index) {
-        match = nextMatch;
-        match.type = 'comment';
-      }
-    }
-    nextMatch = str.match(blockHelperStartRE);
-    if (nextMatch) {
-      if (!match || nextMatch.index < match.index) {
-        match = nextMatch;
-        match.type = 'block';
-      }
-    }
-    nextMatch = str.match(funcHelperRE);
-    if (nextMatch) {
-      if (!match || nextMatch.index < match.index) {
-        match = nextMatch;
-        match.type = 'func';
-      }
-    }
-    if (!match) {
-      break;
-    }
-    str = str.substring(match.index + match[0].length);
-
-    // Translate the match into an appropriate chunk of javascript.
-    if (match.type === 'comment') {
-      // Template comment => javascript comment
-      match[1].split("\n").forEach(addComment);
-    } else if (match.type === 'block') {
-      // Template block helper => javascript function call
-      var helperName = match[1];
-      buf.push(helperName);
-      buf.push('("');
-      var endMatch = str.match(blockHelperEndRE);
-      while (endMatch && endMatch[1] !== helperName) {
-        var skipTo = endMatch.index + endMatch[0].length;
-        buf.push(str.substring(0, skipTo).replace('"', '\\"'));
-        str = str.substring(skipTo);
-        endMatch = str.match(blockHelperEndRE);
-      }
-      if (endMatch) {
-        buf.push(str.substring(0, endMatch.index).replace('"', '\\"'));
-        str = str.substring(endMatch.index + endMatch[0].length);
-      } else {
-        buf.push(str.replace('"', '\\"'));
-        str = '';
-      }
-      buf.push('")\n');
-    } else if (match.type === 'func') {
-      // Template function helper => javascript function call
-      buf.push(match[1]);
-      buf.push('(');
-      buf.push(match[2]);
-      buf.push(')\n');
-    }
-  }
-
-  return buf.join('');
-}
-
-// generate extracted strings file from Handlebars/Mustache templates
-function genHandlebars(hbSources, options) {
-  Object.keys(hbSources).forEach(function (filename) {
-    hbSources[filename] = parseHandlebars(hbSources[filename]);
-  });
-  return gen(hbSources, options);
-}
-
-// generate extracted strings file from Jinja2 templates
-function genJinja(jinjaSources, options) {
-  Object.keys(jinjaSources).forEach(function (filename) {
-    jinjaSources[filename] = parseEJS(jinjaSources[filename], {open: "{{", close: "}}"});
-  });
-
-  return gen(jinjaSources, options);
-}
-
-exports.generate               = gen;
-exports.generateFromEJS        = genEJS;
-exports.generateFromJade       = genJade;
-exports.generateFromHandlebars = genHandlebars;
-exports.generateFromJinja      = genJinja;
+exports.generate = gen;
