+++ tests/ejs.js
@@ -2,16 +2,17 @@
 
 var fs = require('fs');
 var jsxgettext = require('../lib/jsxgettext');
+var ejs = require('../lib/parsers/ejs').ejs;
 var path = require('path');
 
 exports['test ejs'] = function (assert, cb) {
   // check that include syntax doesn't break extraction
   var inputFilename = path.join(__dirname, 'inputs', 'include.ejs');
   fs.readFile(inputFilename, "utf8", function (err, source) {
+    var result = jsxgettext.generate.apply(jsxgettext, ejs(
+      {'inputs/include.ejs': source}, {})
+    );
 
-    var opts = {},
-        sources = {'inputs/include.ejs': source},
-        result = jsxgettext.generateFromEJS(sources, 'inputs/include.js', opts);
     assert.equal(typeof result, 'string', 'result is a string');
     assert.ok(result.length > 1, 'result is not empty');
     assert.ok(result.indexOf('this is a localizable string') !== -1,
@@ -19,6 +20,5 @@ exports['test ejs'] = function (assert, cb) {
     cb();
   });
 };
-
 
 if (module === require.main) require('test').run(exports);
